refactor(icemodel): Simplify substep logical flags, replace subfail with subiter.
These are minor logical improvements identified on close inspection.

No impact on model structure or performance.

The "OK==false" check is not needed on the while loop over substeps, it controls the RESETSUBSTEP call.

The subfail<maxsubiter forced RESETSUBSTEP to be called every time OK is false, including when dt_new == dt_min.

Had to add dt_new as an input to NEXTSTEP in case OK is false.

That never happened in icemodel (good) but happened when testing skimodel with a convergence flag.

That indicates cases where dt_new == dt_min and OK is stuck on false.