fix(UPDATESUBSTEP): Remove max(dt-dt_sum, dt_min) so dt_FULL_STEP is exactly completed.
In prior method, if dt_new=dt_min, the max(...) condition disallowed shortening dt_new to the exact amount to complete the step.

Fix TINY conditions to guarantee correct calculation within tolerance.

Remove dt_flag returned by UPDATESUBSTEP.

Remove dt_flag initialization and its use in NEXTSTEP.

dt_flag had the intended effect on the second-to-last substep but as designed was subsequently set false on the final one and thus had no effect on NEXTSTEP.

It is not necessary, the reset in NEXTSTEP works fine without it.