feat(SEB): Add QADVECT and Qa term to SEBSOLVE.
Refactor SFCTEMP to implicitly handle non-zero Qc derivative.

For Patankar Practice B, Qc is into the skin; for Practice A it is in to the top CV.

Thus for Practice B, Qc = f(Ts), but Qc could also be considered a known constant.

Either way, one must decide if -a(1) is included in dSEB/dTs.

SFCTEMP was redesigned to implicitly include a non-zero -a(1) if an input flag is true.

opts.conduct_type option differentiates Patankar boundary node Practice A from B in CONDUCT.